log_generator/requirements.txt
Faker==0.8.7
pylint==1.8.1
tzlocal==1.5.1
kafka-python==1.4.7